Extract Profile trait
